Fix trail issues
- trail no longer flickers
- trail grows smoothly instead of discretly

First, the clone created by the LeaveTrailComponent is turned fully red with it's own shader. Since the FullTrailMaskGen Viewport is set to never update, the sprite leaves a red trail that is turned white by RedToWhiteShader, that sits under all the clones. After that the ShrinkShader turns all the white pixels neighboring black pixels from top a little darker. Then the generated texture gets sent to the TrailMaskGen, where it is scaled up and cropped, with a margin. After that MosaicShader applies Worley Mosaic to this Viewport. Finally the texture is sent to the ground mask, where it is cropped, and the red value of pixels is turned into an alpha value for the Ground node.