Start work on worm v3
The solution I came up with for Worm was convoluted, because I had no idea what exactly I was doing. The new version will be much simpler: most information will be stored in the Head Node, as it should be. The segments will be unaware about controls, and all communication from them to the head will be passed with signals.